fix(middlew): move the middleware file
